third_party/blink/web_tests/battery-status/multiple-promises-after-resolve.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/services/device/public/mojom/battery_monitor.mojom.js"></script>
<script src="resources/mock-battery-monitor.js"></script>
<script>
promise_test(async _ => {
  const promise1 = navigator.getBattery();
  setAndFireMockBatteryInfo(false, 10, 20, 0.5);
  const result1 = await promise1;
  testIfBatteryStatusIsUpToDate(result1);
  const promise2 = navigator.getBattery();
  assert_equals(promise1, promise2);
  const result2 = await promise2;
  testIfBatteryStatusIsUpToDate(result2);
}, 'multiple consecutive invocations of navigator.getBattery()');
</script>
</body>
</html>
